requirements.txt
#django
django==2.1.10
django-environ==0.4.5
django-extensions==2.1.0
djangorestframework==3.9.1
django-rest-auth==0.3.3
drf-nested-routers==0.90.2
django-filter==2.0.0
requests==2.20.0
django-cors-headers==2.4.0
django-constance==2.3.1
django-picklefield==2.0
whitenoise==3.3.1
django-debug-toolbar==1.7

#DB
psycopg2==2.7.5
psycopg2-binary==2.7.5

#Code Style
pre-commit==1.11.0
black==18.6b4

#Docs
drf-yasg==1.10.2

#Code Style
black==18.6b4

# WSGI Handler
gevent==1.2.1
gunicorn==19.7.1
